Use a generic runner class
