+++ src/kompos/__init__.py
@@ -12,7 +12,7 @@ from subprocess import call
 
 from termcolor import colored
 
-__version__ = "0.4.0-local"
+__version__ = "0.4.0"
 
 
 def display(msg, color):
