+++ src/kompos/komposconfig.py
@@ -12,7 +12,6 @@ import logging
 import os
 from distutils.version import StrictVersion
 from functools import reduce
-from pathlib import Path
 
 import fastjsonschema
 import yaml
@@ -22,27 +21,6 @@ from kompos import __version__
 logger = logging.getLogger(__name__)
 
 CONFIG_SCHEMA_PATH = "data/config_schema.json"
-
-# The filename of the generated hierarchical configuration for Terrraform.
-TERRAFORM_CONFIG_FILENAME = "variables.tfvars.json"
-
-# The filename of the generated Terrraform provider.
-TERRAFORM_PROVIDER_FILENAME = "provider.tf.json"
-
-# The filename of the generated hierarchical configuration for Helmfile.
-HELMFILE_CONFIG_FILENAME = "generated-values.yaml"
-
-# Directory to store terraform plugin cache
-TERRAFORM_CACHE_DIR = "~/.kompos/.terraform.d/plugin-cache"
-
-
-def local_config_dir(directory=TERRAFORM_CACHE_DIR):
-    try:
-        Path(Path.expanduser(Path(directory))).mkdir(parents=True, exist_ok=True)
-        return Path.expanduser(Path(directory))
-
-    except IOError:
-        logging.error("Failed to create dir in path: %s", directory)
 
 
 def get_value_or(dictionary, x_path, default=None):
